# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pulsar_data_aggregation_API",
        "description": "",
        "contact": {
            "email": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/suggestions-category/get-by-id/{id}": {
            "get": {
                "tags": [
                    "Suggestions Category"
                ],
                "operationId": "b5fdb61c82f7c310f85508168ed26135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/suggestions-category": {
            "post": {
                "tags": [
                    "Suggestions Category"
                ],
                "operationId": "45197e466aeff93ba07f4c6c1f5a6c02",
                "requestBody": {
                    "request": "Create Suggestions Category",
                    "description": "Create Suggestions Category Fields",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "0",
                                        "type": "string",
                                        "example": "Текст"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "name": {
                                            "description": "0",
                                            "type": "string",
                                            "example": "Текст"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "errors",
                                    "description": "errors object",
                                    "properties": {
                                        "errors": {
                                            "title": "Validation errors",
                                            "description": "Validation errors object",
                                            "properties": {
                                                "field1": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field1 field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/suggestions-category/{id}": {
            "delete": {
                "tags": [
                    "Suggestions Category"
                ],
                "operationId": "a9e03e339707cba4539d030804c8d56a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_deleted": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error Delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_deleted": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Suggestions Category"
                ],
                "operationId": "50dfbed5b4eb48e8920d58010fd1accc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "request": "Update Suggestions Category",
                    "description": "Update Suggestions Category Fields",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "0",
                                        "type": "string",
                                        "example": "Текст"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "name": {
                                            "description": "0",
                                            "type": "string",
                                            "example": "Текст"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "errors",
                                    "description": "errors object",
                                    "properties": {
                                        "errors": {
                                            "title": "Validation errors",
                                            "description": "Validation errors object",
                                            "properties": {
                                                "field1": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field1 field is required."
                                                    }
                                                },
                                                "field2": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field2 field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}